save fuckfuck

+++ libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: swautele <swautele@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/10 15:43:13 by swautele          #+#    #+#             */
-/*   Updated: 2022/05/12 13:04:51 by swautele         ###   ########.fr       */
+/*   Updated: 2022/05/12 13:09:49 by swautele         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ char	*ft_strnstr(const char *str, const char *searched, size_t n)
 		return ((char *)str);
 	if (n == 0)
 		return (NULL);
-	while (*str && (ssize_t)n - (ssize_t)ft_strlen(searched) >= 0)
+	while (*str && n - ft_strlen(searched) >= 0)
 	{
 		i = 0;
 		while (str[i] == searched[i] && searched[i] && n - i > 0)
